+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.4.0] 2024-06-12
+
 ### Added
 
 - Now supports vite development with React.